move Sv1ServerChannelState to sv1_server
